+++ src/main/java/service/ShipperService.java
@@ -8,7 +8,7 @@ import java.util.List;
  * Created by dengrong on 2016/1/3.
  */
 public interface ShipperService {
-    public boolean evaluateOrder(int userId, int id, int evaluateType, String content);
+    public boolean evaluateOrder(int userId, int id, Byte evaluateType, String content);
     public List getFrequentDrivers(int userId);
     public List getFrequentAddresses(int userId);
     public List searchDriversByPhoneNum(String phoneNum);
